anonym identity logic
